arbiter: Remove SIG_QUEUE and only use PIPE instead
Since memory allocations aren't recommended when we're in signal
handling context, it's most likely unsafe to manipulate a python list
inside such context.

Instead, when we're in a signal handler, we write the signal number to
a pipe, and then we read the pipe once we've exited the signal
context. This way, we won't to any memory allocation in a signal
context, as the handling takes place in the "normal" context.

So after this commit, PIPE is used both for waking up the arbiter
process (prematurely before the timeout has been reached), and for
sending what incoming signal we've received.

This pipe can also be used when terminating gunicorn: Since we know
when a child terminates by just reading the pipe, we don't need to
unconditionally sleep 0.1 seconds -- we just wait for for an event.